Added a save config
I still need to iron it out and make it useable, but it does save an ini file now with all of the correct information in it. There are some function calls I need to get in order though as they are either out of place or missing in some places now.